fix: resolve all compilation errors for v0.3.100 (final)

+++ apps/mobile/lib/features/reports/data/income_statement_provider.dart
@@ -236,9 +236,9 @@ final incomeStatementMonthlyTrendProvider = FutureProvider<List<MonthlyIncomeSta
 
   return monthlyData.map((data) => MonthlyIncomeStatementData(
     monthLabel: data.monthLabel,
-    revenue: Decimal.fromInt(data.revenueNum) / Decimal.fromInt(data.denom),
-    expenses: Decimal.fromInt(data.expenseNum) / Decimal.fromInt(data.denom),
-    netIncome: Decimal.fromInt(data.netIncomeNum) / Decimal.fromInt(data.denom),
+    revenue: (Decimal.fromInt(data.revenueNum) / Decimal.fromInt(data.denom)).toDecimal(),
+    expenses: (Decimal.fromInt(data.expenseNum) / Decimal.fromInt(data.denom)).toDecimal(),
+    netIncome: (Decimal.fromInt(data.netIncomeNum) / Decimal.fromInt(data.denom)).toDecimal(),
   )).toList();
 });
